requirements.txt
langchain
openai
streamlit
tiktoken
python-dotenv
duckduckgo-search